feat: add confirmation state
