+++ src/components/Order/components/send_order.tsx
@@ -4,7 +4,7 @@ import { toast } from 'react-toastify';
 import {
   useDeleteCartMutation,
   useGetActiveCartQuery,
-} from '../../services/api';
+} from '../../../services/api';
 
 const SendOrder = ({ onCartDeleted }: { onCartDeleted: () => void }) => {
   const { data: cart } = useGetActiveCartQuery({});
